requirements.txt
Django==4.2.1
djangorestframework==3.14.0
Faker==18.9.0
xhtml2pdf==0.2.11